Move file util functions
