+++ cmd/render.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/kartverket/skipctl/pkg/constants"
 	"github.com/kartverket/skipctl/pkg/manifest"
+	fileutil "github.com/kartverket/skipctl/pkg/util"
 	"github.com/spf13/cobra"
 )
 
@@ -29,7 +30,7 @@ correctly, otherwise return code 1 is used to indicate failure.`,
 }
 
 func runRender(_ *cobra.Command, _ []string) {
-	files, err := findFilesWithSuffixes(path, constants.ManifestSuffixes)
+	files, err := fileutil.FindFilesWithSuffixes(path, constants.ManifestSuffixes)
 
 	if err != nil {
 		log.Error("Error collecting files",
